requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
prometheus-fastapi-instrumentator>=6.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.40b0
opentelemetry-exporter-otlp>=1.20.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
